BitSet compilation + prefsums

+++ numbers/numbers.hpp
@@ -9,7 +9,7 @@ std::vector<T> impl__factorials(std::size_t maxN, const T& one) {
 	std::vector<T> res(maxN + 1);
 	res[0] = one;
 	for (std::size_t i = 1; i <= maxN; ++i) {
-		res[i] = res[i - 1] * i;
+		res[i] = res[i - 1] * static_cast<long long>(i);
 	}
 	return res;
 }
